[Test] Consolidate distributed contract coverage
